refactor: extract context initialization to padz::init
