+++ src/padz/init.rs
@@ -1,0 +1,43 @@
+use crate::api::{PadzApi, PadzPaths};
+use crate::config::PadzConfig;
+use crate::model::Scope;
+use crate::store::fs::FileStore;
+use directories::ProjectDirs;
+use std::path::Path;
+
+pub struct PadzContext {
+    pub api: PadzApi<FileStore>,
+    pub scope: Scope,
+    pub config: PadzConfig,
+}
+
+pub fn initialize(cwd: &Path, use_global: bool) -> PadzContext {
+    let project_padz_dir = cwd.join(".padz");
+
+    let proj_dirs =
+        ProjectDirs::from("com", "padz", "padz").expect("Could not determine config dir");
+    let global_data_dir = proj_dirs.data_dir().to_path_buf();
+
+    let scope = if use_global {
+        Scope::Global
+    } else {
+        Scope::Project
+    };
+
+    let config_dir = match scope {
+        Scope::Project => &project_padz_dir,
+        Scope::Global => &global_data_dir,
+    };
+    let config = PadzConfig::load(config_dir).unwrap_or_default();
+    let file_ext = config.get_file_ext().to_string();
+
+    let store = FileStore::new(Some(project_padz_dir.clone()), global_data_dir.clone())
+        .with_file_ext(&file_ext);
+    let paths = PadzPaths {
+        project: Some(project_padz_dir),
+        global: global_data_dir,
+    };
+    let api = PadzApi::new(store, paths);
+
+    PadzContext { api, scope, config }
+}
